requirements.txt
django
pytz
Pillow
argon2-cffi
django-redis-cache
whitenoise
django_redis
redis
channels
channels_redis
gevent
celery
django-session-security
django-cors-headers